Server permissions (#54)
* add get all tripPermissions

* get userId as query.param at all tripRoutes

* fix bug: Manager chnt delete or change Manager user

* add to userRoute search user by email

+++ server/src/routes/trip/tripService.ts
@@ -197,7 +197,7 @@ export class TripService {
     if (existingTripUser) {
       throw new Error(`User already has permission: ${existingTripUser.permission_level}`);
     }
-  
+
     // Add the new permission
     return await this.tripUserDal.createTripUser(trip, user.id, permission_level);
   }
@@ -218,26 +218,45 @@ export class TripService {
     if (!tripUser) {
       throw new Error("This user dont have permissions for this trip");
     }
+
+    if(tripUser.permission_level === "Manager"){
+      throw new Error("Cant change Manger permission");
+    }
+
     const result = await this.tripUserDal.updateUserPermission(tripUser, newPermission);
     
     return result;
   }
 
   // Delete user permission for a specific trip
-async deleteUserPermission(trip_id: string, user_email: string) {
-  // Validate user existence
-  const user = await this.userDal.getUserByEmail(user_email);
-  if (!user) {
-    throw new Error("User not found is the system");
-  }
-  const tripUser = await this.tripUserDal.findTripUser(trip_id, user.id);
-  if (!tripUser) {
-    throw new Error("This user dont have permissions for this trip");
-  }
-  const result = await this.tripUserDal.deleteUserPermission(tripUser);
+  async deleteUserPermission(trip_id: string, user_email: string) {
+    // Validate user existence
+    const user = await this.userDal.getUserByEmail(user_email);
+    if (!user) {
+      throw new Error("User not found is the system");
+    }
+    const tripUser = await this.tripUserDal.findTripUser(trip_id, user.id);
+    if (!tripUser) {
+      throw new Error("This user dont have permissions for this trip");
+    }
 
-  return result;
-}
+    if(tripUser.permission_level === "Manager"){
+      throw new Error("Cant delete Manger user");
+    }
 
-  
+    const result = await this.tripUserDal.deleteUserPermission(tripUser);
+
+    return result;
+  }
+
+  async getAllUserIdPermissions(trip_id: string) {
+    // Validate user existence
+    const users = await this.tripUserDal.findTripUsersAndPermissions(trip_id);
+    if (!users) {
+      throw new Error("Authorized users not found for this trip");
+    }
+    return users;
+  }
+
+
 }
